Refactor and delete unused files for readability
