+++ MovieHub/src/constants/categories.js
@@ -1,0 +1,14 @@
+export const CATEGORY_TABS = [
+    { id: 'popular', label: 'Popular' },
+    { id: 'top_rated', label: 'Top Rated' },
+    { id: 'upcoming', label: 'Upcoming' },
+    { id: 'now_playing', label: 'Now Playing' }
+];
+
+export const CATEGORY_TITLES = {
+    popular: 'Popular',
+    top_rated: 'Top Rated',
+    upcoming: 'Upcoming',
+    now_playing: 'Now Playing'
+};
+
