--- MovieHub/src/data/comments/comments.json
@@ -1,3 +1,0 @@
-{
-    "comments": {}
-}
