+++ MovieHub/src/hooks/useIntersectionObserver.js
@@ -1,0 +1,34 @@
+import { useEffect, useRef } from 'react';
+
+export const useIntersectionObserver = (callback, { threshold = 0.1 } = {}) => {
+  const targetRef = useRef(null);
+  const callbackRef = useRef(callback);
+
+  useEffect(() => {
+    callbackRef.current = callback;
+  }, [callback]);
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting) {
+          callbackRef.current();
+        }
+      },
+      { threshold }
+    );
+
+    const currentTarget = targetRef.current;
+    if (currentTarget) {
+      observer.observe(currentTarget);
+    }
+
+    return () => {
+      if (currentTarget) {
+        observer.unobserve(currentTarget);
+      }
+    };
+  }, [threshold]);
+
+  return targetRef;
+};
