+++ MovieHub/src/utils/api.js
@@ -1,0 +1,11 @@
+const API_KEY = import.meta.env.VITE_API_TMDB_KEY;
+
+export const API_BASE_URL = 'https://api.themoviedb.org/3';
+
+export const API_OPTIONS = {
+  method: 'GET',
+  headers: {
+    accept: 'application/json',
+    Authorization: `Bearer ${API_KEY}`,
+  },
+};
